+++ src/app/_modules/report/view/dashboard/dashboard.component.ts
@@ -3,6 +3,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { ReportService } from '../../_service/report.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { SharedServiceService } from '@modules/shared/_services/shared-service.service';
+import { GlobaldialogService } from '@services/globaldialog.service';
 @Component({
   selector: 'pros-dashboard',
   templateUrl: './dashboard.component.html',
@@ -23,7 +24,8 @@ export class DashboardComponent implements OnInit {
     public reportService: ReportService,
     private snackbar: MatSnackBar,
     private sharedService: SharedServiceService,
-    private router: Router
+    private router: Router,
+    private globalDialogService: GlobaldialogService
   ) { }
 
   ngOnInit(): void {
@@ -56,12 +58,16 @@ export class DashboardComponent implements OnInit {
   }
 
   delete() {
-    this.reportService.deleteReport((this.reportId.toString())).subscribe(res=>{
-      if(res) {
-        this.sharedService.setReportListData();
-        this.snackbar.open(`Successfully Deleted`, 'Close',{duration:3000});
+    this.globalDialogService.confirm({label:'Are you sure to delete ?'}, (response) =>{
+      if(response && response === 'yes') {
+        this.reportService.deleteReport((this.reportId.toString())).subscribe(res=>{
+          if(res) {
+            this.sharedService.setReportListData();
+            this.snackbar.open(`Successfully Deleted`, 'Close',{duration:3000});
+          }
+        },err=>console.error(`Error: ${err}`))
       }
-    },err=>console.error(`Error: ${err}`))
+    });
   }
 
   editReport() {
